Divide search tree bloomer's filters in different classes
