+++ ra-optimisation/search-tree-rao/src/main/java/com/powsybl/openrao/searchtreerao/searchtree/algorithms/FarFromMostLimitingElementFilter.java
@@ -1,0 +1,91 @@
+/*
+ * Copyright (c) 2024, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.powsybl.openrao.searchtreerao.searchtree.algorithms;
+
+import com.powsybl.iidm.network.Country;
+import com.powsybl.iidm.network.Network;
+import com.powsybl.openrao.commons.CountryGraph;
+import com.powsybl.openrao.data.cracapi.networkaction.NetworkAction;
+import com.powsybl.openrao.searchtreerao.commons.NetworkActionCombination;
+
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import static com.powsybl.openrao.commons.logs.OpenRaoLoggerProvider.TECHNICAL_LOGS;
+
+/**
+ * @author Thomas Bouquet {@literal <thomas.bouquet at rte-france.com>}
+ */
+public class FarFromMostLimitingElementFilter implements NetworkActionCombinationFilter {
+    private final Network network;
+    private final CountryGraph countryGraph;
+    private final boolean filterFarElements;
+    private final int maxNumberOfBoundariesForSkippingNetworkActions;
+
+    public FarFromMostLimitingElementFilter(Network network, boolean filterFarElements, int maxNumberOfBoundariesForSkippingNetworkActions) {
+        this.network = network;
+        countryGraph = new CountryGraph(network);
+        this.filterFarElements = filterFarElements;
+        this.maxNumberOfBoundariesForSkippingNetworkActions = maxNumberOfBoundariesForSkippingNetworkActions;
+    }
+
+    /**
+     * Removes network actions far from most limiting elements, using the user's parameters for activating/deactivating this
+     * feature, and setting the number of boundaries allowed between the network action and the limiting element.
+     * The most limiting elements are the most limiting functional cost element, and all elements with a non-zero virtual cost.
+     */
+    public Map<NetworkActionCombination, Boolean> filter(Map<NetworkActionCombination, Boolean> naCombinations, Leaf fromLeaf) {
+        if (!filterFarElements) {
+            return naCombinations;
+        }
+
+        Set<Optional<Country>> worstCnecLocation = getOptimizedMostLimitingElementsLocation(fromLeaf);
+
+        Map<NetworkActionCombination, Boolean> filteredNaCombinations = naCombinations.keySet().stream()
+            .filter(naCombination -> naCombination.getNetworkActionSet().stream().anyMatch(na -> isNetworkActionCloseToLocations(na, worstCnecLocation, countryGraph)))
+            .collect(Collectors.toMap(naCombination -> naCombination, naCombinations::get));
+
+        if (naCombinations.size() > filteredNaCombinations.size()) {
+            TECHNICAL_LOGS.info("{} network action combinations have been filtered out because they are too far from the most limiting element", naCombinations.size() - filteredNaCombinations.size());
+        }
+        return filteredNaCombinations;
+    }
+
+    Set<Optional<Country>> getOptimizedMostLimitingElementsLocation(Leaf leaf) {
+        Set<Optional<Country>> locations = new HashSet<>();
+        leaf.getMostLimitingElements(1).forEach(element -> locations.addAll(element.getLocation(network)));
+        for (String virtualCost : leaf.getVirtualCostNames()) {
+            leaf.getCostlyElements(virtualCost, Integer.MAX_VALUE).forEach(element -> locations.addAll(element.getLocation(network)));
+        }
+        return locations;
+    }
+
+    /**
+     * Says if a network action is close to a given set of countries, respecting the maximum number of boundaries
+     */
+    boolean isNetworkActionCloseToLocations(NetworkAction networkAction, Set<Optional<Country>> locations, CountryGraph countryGraph) {
+        if (locations.stream().anyMatch(Optional::isEmpty)) {
+            return true;
+        }
+        Set<Optional<Country>> networkActionCountries = networkAction.getLocation(network);
+        if (networkActionCountries.stream().anyMatch(Optional::isEmpty)) {
+            return true;
+        }
+        for (Optional<Country> location : locations) {
+            for (Optional<Country> networkActionCountry : networkActionCountries) {
+                if (location.isPresent() && networkActionCountry.isPresent()
+                    && countryGraph.areNeighbors(location.get(), networkActionCountry.get(), maxNumberOfBoundariesForSkippingNetworkActions)) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+}
