+++ ra-optimisation/search-tree-rao/src/test/java/com/powsybl/openrao/searchtreerao/searchtree/algorithms/FarFromMostLimitingElementFilterTest.java
@@ -1,0 +1,156 @@
+/*
+ * Copyright (c) 2024, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.powsybl.openrao.searchtreerao.searchtree.algorithms;
+
+import com.powsybl.iidm.network.Country;
+import com.powsybl.openrao.commons.CountryBoundary;
+import com.powsybl.openrao.commons.CountryGraph;
+import com.powsybl.openrao.data.cracapi.networkaction.ActionType;
+import com.powsybl.openrao.data.cracapi.networkaction.NetworkAction;
+import com.powsybl.openrao.data.cracapi.usagerule.UsageMethod;
+import com.powsybl.openrao.searchtreerao.commons.NetworkActionCombination;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
+
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.PREVENTIVE_INSTANT_ID;
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.comb2BeNl;
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.comb2De;
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.comb2FrDeBe;
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.comb3Be;
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.indBe1;
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.indDe1;
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.indDeNl;
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.indFr2;
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.indFrDe;
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.indNl1;
+import static com.powsybl.openrao.searchtreerao.searchtree.algorithms.NetworkActionCombinationsUtils.indNlBe;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.anyInt;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.mock;
+
+/**
+ * @author Thomas Bouquet {@literal <thomas.bouquet at rte-france.com>}
+ */
+class FarFromMostLimitingElementFilterTest {
+    @Test
+    void testRemoveNetworkActionsFarFromMostLimitingElement() {
+        // arrange naCombination list
+        List<NetworkActionCombination> listOfNaCombinations = List.of(indFr2, indDe1, indBe1, indNl1, indNlBe, indFrDe, indDeNl, comb3Be, comb2De, comb2FrDeBe, comb2BeNl);
+        Map<NetworkActionCombination, Boolean> naCombinations = new HashMap<>();
+        listOfNaCombinations.forEach(na -> naCombinations.put(na, false));
+
+        // arrange previous Leaf -> most limiting element is in DE/FR
+        Leaf previousLeaf = mock(Leaf.class);
+        Mockito.when(previousLeaf.getVirtualCostNames()).thenReturn(Collections.emptySet());
+
+        FarFromMostLimitingElementFilter naFilter;
+        Map<NetworkActionCombination, Boolean> filteredNaCombination;
+
+        // test - no border cross, most limiting element is in BE/FR
+        Mockito.when(previousLeaf.getMostLimitingElements(1)).thenReturn(List.of(NetworkActionCombinationsUtils.crac.getFlowCnec("cnec1basecase"))); // be fr
+        naFilter = new FarFromMostLimitingElementFilter(NetworkActionCombinationsUtils.network, true, 0);
+        filteredNaCombination = naFilter.filter(naCombinations, previousLeaf);
+
+        assertEquals(7, filteredNaCombination.size());
+        List<NetworkActionCombination> list1 = List.of(indFr2, indBe1, indNlBe, indFrDe, comb3Be, comb2FrDeBe, comb2BeNl);
+        Map<NetworkActionCombination, Boolean> finalFilteredNaCombination = filteredNaCombination;
+        list1.forEach(na -> assertTrue(finalFilteredNaCombination.containsKey(na)));
+
+        // test - no border cross, most limiting element is in DE/FR
+        Mockito.when(previousLeaf.getMostLimitingElements(1)).thenReturn(List.of(NetworkActionCombinationsUtils.crac.getFlowCnec("cnec2basecase"))); // de fr
+        filteredNaCombination = naFilter.filter(naCombinations, previousLeaf);
+
+        assertEquals(6, filteredNaCombination.size());
+        List<NetworkActionCombination> list2 = List.of(indFr2, indDe1, indFrDe, indDeNl, comb2De, comb2FrDeBe);
+        Map<NetworkActionCombination, Boolean> finalFilteredNaCombination2 = filteredNaCombination;
+        list2.forEach(na -> assertTrue(finalFilteredNaCombination2.containsKey(na)));
+
+        // test - max 1 border cross, most limiting element is in BE
+        Mockito.when(previousLeaf.getMostLimitingElements(1)).thenReturn(List.of(NetworkActionCombinationsUtils.crac.getFlowCnec("cnecBe"))); // be
+        naFilter = new FarFromMostLimitingElementFilter(NetworkActionCombinationsUtils.network, true, 1);
+        filteredNaCombination = naFilter.filter(naCombinations, previousLeaf);
+
+        assertEquals(9, filteredNaCombination.size());
+        List<NetworkActionCombination> list3 = List.of(indFr2, indBe1, indNl1, indNlBe, indFrDe, indDeNl, comb3Be, comb2FrDeBe, comb2BeNl);
+        Map<NetworkActionCombination, Boolean> finalFilteredNaCombination3 = filteredNaCombination;
+        list3.forEach(na -> assertTrue(finalFilteredNaCombination3.containsKey(na)));
+    }
+
+    @Test
+    void testGetOptimizedMostLimitingElementsLocation() {
+        FarFromMostLimitingElementFilter naFilter = new FarFromMostLimitingElementFilter(NetworkActionCombinationsUtils.network, false, 0);
+
+        Leaf leaf = mock(Leaf.class);
+        Mockito.when(leaf.getVirtualCostNames()).thenReturn(Set.of("mnec", "lf"));
+
+        Mockito.when(leaf.getMostLimitingElements(1)).thenReturn(List.of(NetworkActionCombinationsUtils.crac.getFlowCnec("cnec1basecase"))); // be fr
+        Mockito.when(leaf.getCostlyElements(eq("mnec"), anyInt())).thenReturn(List.of(NetworkActionCombinationsUtils.crac.getFlowCnec("cnec2basecase"))); // de fr
+        Mockito.when(leaf.getCostlyElements(eq("lf"), anyInt())).thenReturn(Collections.emptyList());
+        assertEquals(Set.of(Optional.of(Country.BE), Optional.of(Country.FR), Optional.of(Country.DE)), naFilter.getOptimizedMostLimitingElementsLocation(leaf));
+
+        Mockito.when(leaf.getMostLimitingElements(1)).thenReturn(List.of(NetworkActionCombinationsUtils.crac.getFlowCnec("cnec1basecase"))); // be fr
+        Mockito.when(leaf.getCostlyElements(eq("mnec"), anyInt())).thenReturn(Collections.emptyList());
+        Mockito.when(leaf.getCostlyElements(eq("lf"), anyInt())).thenReturn(Collections.emptyList());
+        assertEquals(Set.of(Optional.of(Country.BE), Optional.of(Country.FR)), naFilter.getOptimizedMostLimitingElementsLocation(leaf));
+
+        Mockito.when(leaf.getMostLimitingElements(1)).thenReturn(Collections.emptyList());
+        Mockito.when(leaf.getCostlyElements(eq("mnec"), anyInt())).thenReturn(Collections.emptyList());
+        Mockito.when(leaf.getCostlyElements(eq("lf"), anyInt())).thenReturn(List.of(NetworkActionCombinationsUtils.crac.getFlowCnec("cnec2basecase"))); // de fr
+        assertEquals(Set.of(Optional.of(Country.FR), Optional.of(Country.DE)), naFilter.getOptimizedMostLimitingElementsLocation(leaf));
+
+        Mockito.when(leaf.getMostLimitingElements(1)).thenReturn(Collections.emptyList());
+        Mockito.when(leaf.getCostlyElements(eq("mnec"), anyInt())).thenReturn(List.of(NetworkActionCombinationsUtils.crac.getFlowCnec("cnec1basecase"), NetworkActionCombinationsUtils.crac.getFlowCnec("cnec2basecase"))); // be de fr
+        Mockito.when(leaf.getCostlyElements(eq("lf"), anyInt())).thenReturn(Collections.emptyList());
+        assertEquals(Set.of(Optional.of(Country.BE), Optional.of(Country.FR), Optional.of(Country.DE)), naFilter.getOptimizedMostLimitingElementsLocation(leaf));
+
+        Mockito.when(leaf.getMostLimitingElements(1)).thenReturn(Collections.emptyList());
+        Mockito.when(leaf.getCostlyElements(eq("mnec"), anyInt())).thenReturn(List.of(NetworkActionCombinationsUtils.crac.getFlowCnec("cnec2basecase")));
+        Mockito.when(leaf.getCostlyElements(eq("lf"), anyInt())).thenReturn(List.of(NetworkActionCombinationsUtils.crac.getFlowCnec("cnec1basecase")));
+        assertEquals(Set.of(Optional.of(Country.BE), Optional.of(Country.FR), Optional.of(Country.DE)), naFilter.getOptimizedMostLimitingElementsLocation(leaf));
+    }
+
+    @Test
+    void testIsNetworkActionCloseToLocations() {
+        NetworkAction na1 = NetworkActionCombinationsUtils.crac.newNetworkAction().withId("na").newTopologicalAction().withNetworkElement("BBE2AA1  FFR3AA1  1").withActionType(ActionType.OPEN).add().newOnInstantUsageRule().withUsageMethod(UsageMethod.AVAILABLE).withInstant(PREVENTIVE_INSTANT_ID).add().add();
+        NetworkAction na2 = mock(NetworkAction.class);
+        Mockito.when(na2.getLocation(NetworkActionCombinationsUtils.network)).thenReturn(Set.of(Optional.of(Country.FR), Optional.empty()));
+
+        HashSet<CountryBoundary> boundaries = new HashSet<>();
+        boundaries.add(new CountryBoundary(Country.FR, Country.BE));
+        boundaries.add(new CountryBoundary(Country.FR, Country.DE));
+        boundaries.add(new CountryBoundary(Country.DE, Country.AT));
+        CountryGraph countryGraph = new CountryGraph(boundaries);
+
+        FarFromMostLimitingElementFilter naFilter;
+
+        naFilter = new FarFromMostLimitingElementFilter(NetworkActionCombinationsUtils.network, false, 0);
+        assertTrue(naFilter.isNetworkActionCloseToLocations(na1, Set.of(Optional.empty()), countryGraph));
+        assertTrue(naFilter.isNetworkActionCloseToLocations(na1, Set.of(Optional.of(Country.FR)), countryGraph));
+        assertTrue(naFilter.isNetworkActionCloseToLocations(na1, Set.of(Optional.of(Country.BE)), countryGraph));
+        assertFalse(naFilter.isNetworkActionCloseToLocations(na1, Set.of(Optional.of(Country.DE)), countryGraph));
+        assertFalse(naFilter.isNetworkActionCloseToLocations(na1, Set.of(Optional.of(Country.AT)), countryGraph));
+        assertTrue(naFilter.isNetworkActionCloseToLocations(na2, Set.of(Optional.of(Country.AT)), countryGraph));
+
+        naFilter = new FarFromMostLimitingElementFilter(NetworkActionCombinationsUtils.network, true, 1);
+        assertTrue(naFilter.isNetworkActionCloseToLocations(na1, Set.of(Optional.of(Country.DE)), countryGraph));
+        assertFalse(naFilter.isNetworkActionCloseToLocations(na1, Set.of(Optional.of(Country.AT)), countryGraph));
+
+        naFilter = new FarFromMostLimitingElementFilter(NetworkActionCombinationsUtils.network, true, 2);
+        assertTrue(naFilter.isNetworkActionCloseToLocations(na1, Set.of(Optional.of(Country.AT)), countryGraph));
+    }
+}
